start `_blocks` module, move `compute_out_shape` to `_blocks`
